Fixed NASTY problem with some yes/no questions where Meseta were involved.
Turns out the logic checks for a "!" mark as the
last character of the dialogue tree from the script
to check what answer was selected.

This seems rather nasty... basically it means we
need to make sure those positive answers end with
an exclamation mark and the negative one doesn't.

Took quite some time to figure that one out...
When I dumped memory around where stuff was being
checked (and noticed it was the cmpi.b instruction
that was failing in French but not in English), I
realized that the immediate constant was probably
some sort of char comparison.  And true enough, it
was the value for the ! character.  Went from there.