Fix handling for "special" yes-no dialogues.
Those dialogues have some effect on Meseta if
you answer "Yes", on top of dialogue effects.
The code used to check for the last character
of the dialogue being a "!" character. However,
for translation purposes, this is not ideal.

I changed it to use #$00 (actually the space
character in the dialogue window) but made it
an actual constant so it's clearer what's going
on.

Feel free to change it to something else and
munge the dialogue font to make it a blank,
or even a control code of some sort that
doesn't do anything.  A space is pretty
suitable however because the dialogue window
will clip extra characters anyway (unlike the
small message window) and dialogues don't
usually end with a space anyway.

This does waste a byte of space but at this
point, with the whole thing uncompressed
anyway, nobody will care :)